New Feature  & README Update: Changing public leaderboard font
Now there is the ability to change the custom leaderboard font. The user can open the settings page and upload fonts to the site changing the font on the public leaderboard. The inclusion of reset to defaults button was also included allowing the user to reset to the default font.

Also kinda updated readme?